move ScheduleGrid subcomponents into subfolders, extract useScheduleSelection and useScheduleGridData hooks from ScheduleGrid
